DaliDoodle-Code/requirements.txt
brotli==1.0.9
ConfigParser==5.3.0
cryptography==40.0.2
Cython==0.29.28
dl==0.1.0
docutils==0.20.1
en_core_web_sm==3.4.1
Flask==2.3.2
Flask_SocketIO==5.3.3
happytransformer==2.4.1
HTMLParser==0.0.2
importlib_metadata==6.4.1
jnius==1.1.0
keyring==24.2.0
lockfile==0.12.2
lxml==4.9.2
numpy==1.23.5
openai==0.27.6
ordereddict==1.1
Pillow==9.5.0
protobuf==4.23.3
pymongo==4.3.3
pyOpenSSL==23.2.0
pywin32==305
sentence_transformers==2.2.2
spacy==3.4.4
toml==0.10.2
torch==2.0.1+cu118
tornado==6.3.2
transformers==4.26.1
typing_extensions==4.6.3
xmlrpclib==1.0.1
zipp==3.15.0
